requirement.txt
aiosignal==1.2.0
asgiref==3.5.2
backports.zoneinfo==0.2.1
bcrypt==4.0.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
Django==4.0
django-cors-headers==3.13.0
django-countries==7.5
django-dotenv==1.4.2
django-extensions==3.2.1
django-filter==22.1
django-phonenumber-field==6.4.0
django-phonenumbers==1.0.1
django-rest-swagger==2.2.0
django-restframework==0.0.1
django-seed==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-writable-nested==0.7.0
drf-yasg==1.21.3
et-xmlfile==1.1.0
Faker==14.2.0
filelock==3.9.0
frozenlist==1.3.3
graphviz==0.20.1
huggingface-hub==0.9.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.3
openapi-codec==1.3.2
openpyxl==3.0.10
openpyxl-image-loader==1.0.5
packaging==21.3
pandas==1.5.0
paramiko==2.11.0
phonenumbers==8.12.54
Pillow==9.3.0
psycopg2==2.9.3
pycparser==2.21
pydot==1.4.2
PyJWT==2.4.0
PyNaCl==1.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
toposort==1.7
tqdm==4.64.1
typing_extensions==4.4.0
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
